+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/reporting/repository/ReportJdbcTemplate.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.approvedpremisesapi.reporting.util
+package uk.gov.justice.digital.hmpps.approvedpremisesapi.reporting.repository
 
 import org.springframework.jdbc.core.ResultSetExtractor
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate
